requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psutil==7.0.0
Werkzeug==3.1.3
flask-bcrypt==1.0.1
# Pour envoi d’e-mails
# Pour la gestion des fichiers
python-dotenv
requests